deps only

+++ install.sh
@@ -20,6 +20,8 @@ elif [[ "$IS_UBUNTU" == true ]]; then
       exit 1
       ;;
   esac
+elif [[ "$IS_ARCH" == true ]]; then
+  TASKS_DIR="$DOTFILES_DIR/tasks/arch"
 else
   echo "🚨 Unsupported OS/Distribution" >&2
   exit 1
@@ -34,37 +36,39 @@ if [[ -f "$TASKS_DIR/00_prerequisites.sh" ]]; then
   fi
 fi
 
-if [[ "$WORK" == false ]]; then
-  if [[ ! -f "$DOTFILES_DIR/modules/private/vault.yml" ]]; then
-    echo "🚨 Private submodule required"
-    echo "Generate SSH and GPG keys and add them to GitHub"
-    echo "Then run: 'git submodule update --init --recursive'"
-    exit 1
+if [[ "$DEPS_ONLY" == false ]]; then
+  if [[ "$WORK" == false ]]; then
+    if [[ ! -f "$DOTFILES_DIR/modules/private/vault.yml" ]]; then
+      echo "🚨 Private submodule required"
+      echo "Generate SSH and GPG keys and add them to GitHub"
+      echo "Then run: 'git submodule update --init --recursive'"
+      exit 1
+    fi
+
+    read -sp "Vault password: " VAULT_PW
+    echo
+
+    VAULT_PW_FILE=$(mktemp)
+    chmod 600 "$VAULT_PW_FILE"
+    echo "$VAULT_PW" > "$VAULT_PW_FILE"
+
+    ansible-playbook "$DOTFILES_DIR/main.yml" \
+      -e "work_mode=$WORK" \
+      -e "email=$EMAIL" \
+      --vault-password-file "$VAULT_PW_FILE"
+  else
+    ansible-playbook "$DOTFILES_DIR/main.yml" \
+      -e "work_mode=$WORK" \
+      -e "email=$EMAIL"
   fi
-
-  read -sp "Vault password: " VAULT_PW
-  echo
-
-  VAULT_PW_FILE=$(mktemp)
-  chmod 600 "$VAULT_PW_FILE"
-  echo "$VAULT_PW" > "$VAULT_PW_FILE"
-
-  ansible-playbook "$DOTFILES_DIR/main.yml" \
-    -e "work_mode=$WORK" \
-    -e "email=$EMAIL" \
-    --vault-password-file "$VAULT_PW_FILE"
-else
-  ansible-playbook "$DOTFILES_DIR/main.yml" \
-    -e "work_mode=$WORK" \
-    -e "email=$EMAIL"
 fi
-exit 0
+
 for task in "$TASKS_DIR"/*.sh; do
   if [[ -f "$task" ]]; then
     source "$task"
     TASK_NAME=$(basename "${task%.sh}")
 
-    if TASK_NAME == "00_prerequisites"; then
+    if $TASK_NAME == "00_prerequisites"; then
       continue
     fi
